tidy: run clang-tidy where the SDK is (ci.yml), not where it is not (the community pipeline) (#2)
The distribution's Tidy Check failed on main for a reason that was never our code: that job
configures the build without provisioning vcpkg, and this extension cannot configure at all without
the OpenTelemetry SDK - `find_package(opentelemetry-cpp CONFIG REQUIRED)` and a toolchain file that
is not there. So the distribution asks for `format` only, and clang-tidy becomes a step of the Linux
CI job, where vcpkg, the SDK and the compile database already exist: `make tidy` prints findings,
`make tidy-ci` fails the step on one.

Its first run found two, fixed here: a scalar_function_t copied instead of moved, and a `long long`
where duckdb spells int64_t.

+++ src/acl_otel_otlp.cpp
@@ -223,8 +223,7 @@ OtlpConfig OtlpConfig::From(DatabaseInstance &db) {
 #ifndef _WIN32
 		gethostname(host, sizeof(host));
 #endif
-		config.instance_id =
-		    string(host[0] ? host : "node") + ":" + std::to_string(static_cast<long long>(ACL_OTEL_PID));
+		config.instance_id = string(host[0] ? host : "node") + ":" + std::to_string(static_cast<int64_t>(ACL_OTEL_PID));
 	}
 	config.duckdb_version = DuckDB::LibraryVersion();
 #ifdef EXT_VERSION_ACL_OTEL
